Correction

+++ projet.cpp
@@ -109,7 +109,7 @@ Projet* Projet::open(QDir &path)
 
 void Projet::save()
 {
-    for(int i = 0; i < _imagesOutput.size(); i++)
+    for(size_t i = 0; i < _imagesOutput.size(); i++)
     {
         _imagesOutput.at(i)->save(_output->path()+"/img" + QString::number(i) + ".bmp");
     }
